Many minor edits, no answer changes (confirmed by Mac LIVV).
NOTE: Requires a fresh configure and build.

Created a new module, glissade_tests.F90.  This contains a couple of
test subroutines (one for halos, one for transport) that were in glissade.F90.

Fixed a call to glissade_transport from the enthalpy branch of glissade_tstep.
(Optional argument 'waterfrac' was being passed in without a keyword.
Making the fix did not, however, change any results.)

Moved 'ntracer' from glissade.F90 to glissade_transport.F90.

Populated the ubas and vbas arrays at the end of the glissade diagnostic solve
(these were previously filled with zeroes).

Cleaned up more TODO statements (and resolved a few).

Replaced some 'stop' statements with fatal calls to write_log.

Removed some unnecessary variables from argument lists; removed some unused subroutines;
removed some unused variables; fixed some indentation.

git-svn-id: http://oceans11.lanl.gov/svn/PISCEES/trunk@1871 42a2fd4b-9c52-44de-a8a9-1772c67210cd

Former-commit-id: 2b8bf84df67644ee61c6a7b19c5ba29e314b4e74

+++ libglint/glint_downscale.F90
@@ -296,7 +296,7 @@ contains
     ! Set the mbal method and tstep
 
     params%mbal%which = whichacab     ! = 0 for GCM coupling
-    params%mbal%tstep = years2hours   ! no. of hours in 1 year
+    params%mbal%tstep = nint(years2hours)   ! no. of hours in 1 year
 
   end subroutine glint_init_input_gcm
